refactor: create utils.ts
